step_1의 xml과activity를 add_contact로 변경하고 mainActivity생성, mainActivity에서 accContact로 intent생성
