+++ README.md
@@ -1,10 +1,17 @@
 # android-contacts
 
 0. commit 테스트
-1. contact branch 생성
+1. branch [contact_step_1] 생성
 2. xml 1차 작성
-3. MainActivit작성
-   생일 입력 시 캘린더가 나타나는 showDatePickerDialog()
+3. 생일 입력 시 캘린더가 나타나는 showDatePickerDialog()
 4. xml 완성
-5. 이름, 전화번호 입력 확인
-6. 토스트 메시지
+5. 이름, 전화번호 입력 확인 
+6. 토스트 메시지 기능
+
+7. branch [contact_step_2] 생성
+8. 연락처 등록 화면 구현
+9. 연락처를 저장하면 목록에 추가
+10. 저장된 연락처가 많을 경우 스크롤
+11. 추가된 연락처를 선택하여 상세 화면 보기
+12. 연락처 작성 중 뒤로가기 버튼을 누르면 확인 팝업이 나타난다
+13. 추가된 연락처는 앱을 다시 실행하면 유지되지 않는다
