+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,9 @@
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
+            android:exported="false" />
+        <activity
+            android:name=".AddContact"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
