+++ app/src/main/java/campus/tech/kakao/contacts/MainActivity.kt
@@ -1,80 +1,29 @@
 package campus.tech.kakao.contacts
 
-import android.app.DatePickerDialog
+import android.content.Intent
 import android.os.Bundle
-import android.view.View
-import android.widget.*
-
+import android.widget.ImageView
+import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import java.util.*
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 
 class MainActivity : AppCompatActivity() {
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
         setContentView(R.layout.activity_main)
-
-        val toggleButton: ImageView = findViewById(R.id.toggle)
-        val additionalFieldsLayout: LinearLayout = findViewById(R.id.additionalFieldsLayout)
-        val birthdayEditText: EditText = findViewById(R.id.birthdayEditText)
-        val nameEditText: EditText = findViewById(R.id.name)
-        val phoneNumEditText: EditText = findViewById(R.id.phoneNum)
-        val maleRadioButton: RadioButton = findViewById(R.id.maleRadioButton)
-        val femaleRadioButton: RadioButton = findViewById(R.id.femaleRadioButton)
-        val saveButton: TextView = findViewById(R.id.saveButton)
-        val cancelButton: TextView = findViewById(R.id.cancelButton)
-        val toggleLayout: LinearLayout = findViewById(R.id.toggleLayout)
-
-        toggleButton.setOnClickListener {
-            toggleLayout.visibility = View.GONE
-            additionalFieldsLayout.visibility = View.VISIBLE
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
         }
 
-        birthdayEditText.setOnClickListener {
-            showDatePickerDialog()
+        val addContact: ImageView = findViewById(R.id.add_contact)
+
+        addContact.setOnClickListener {
+            val intent = Intent(this@MainActivity, AddContact::class.java)
+            startActivity(intent)
         }
-
-        saveButton.setOnClickListener {
-            val name = nameEditText.text.toString()
-            val phoneNumber = phoneNumEditText.text.toString()
-
-            if (name.isEmpty() || phoneNumber.isEmpty()) {
-                Toast.makeText(this, "이름과 전화번호는 필수 값입니다.", Toast.LENGTH_SHORT).show()
-                return@setOnClickListener
-            }
-
-            if (!phoneNumber.matches("\\d+".toRegex())) {
-                Toast.makeText(this, "전화번호는 숫자만 입력 가능합니다.", Toast.LENGTH_SHORT).show()
-                return@setOnClickListener
-            }
-
-//            if (!maleRadioButton.isChecked && !femaleRadioButton.isChecked) {
-//                Toast.makeText(this, "성별을 선택해 주세요.", Toast.LENGTH_SHORT).show()
-//                return@setOnClickListener
-//            }
-
-            Toast.makeText(this, "저장이 완료되었습니다.", Toast.LENGTH_SHORT).show()
-        }
-
-        cancelButton.setOnClickListener {
-            Toast.makeText(this, "취소 되었습니다.", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    private fun showDatePickerDialog() {
-        val calendar = Calendar.getInstance()
-        val year = calendar.get(Calendar.YEAR)
-        val month = calendar.get(Calendar.MONTH)
-        val day = calendar.get(Calendar.DAY_OF_MONTH)
-
-        val datePickerDialog = DatePickerDialog(this, { _, selectedYear, selectedMonth, selectedDay ->
-            val selectedDate = "$selectedYear-${selectedMonth + 1}-$selectedDay"
-            val birthdayEditText: EditText = findViewById(R.id.birthdayEditText)
-            birthdayEditText.setText(selectedDate)
-        }, year, month, day)
-
-        datePickerDialog.show()
     }
 }
-
-
