+++ app/src/main/res/drawable/add_box_40px.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="60dp"
+    android:height="60dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960"
+    android:tint="?attr/colorControlNormal">
+  <path
+      android:fillColor="#FFFF00"
+      android:pathData="M437.39,689.87L523.46,689.87L523.46,522.61L689.87,522.61L689.87,436.54L523.46,436.54L523.46,270.13L437.39,270.13L437.39,436.54L270.13,436.54L270.13,522.61L437.39,522.61L437.39,689.87ZM189.06,860.2Q152.92,860.2 126.36,833.64Q99.8,807.08 99.8,770.94L99.8,189.06Q99.8,152.85 126.36,126.24Q152.92,99.64 189.06,99.64L770.94,99.64Q807.15,99.64 833.76,126.24Q860.36,152.85 860.36,189.06L860.36,770.94Q860.36,807.08 833.76,833.64Q807.15,860.2 770.94,860.2L189.06,860.2Z"/>
+</vector>
